Fix bug from OpenOrCreate. Update tests. PR.
Using FileMode.OpenOrCreate caused files not to be truncated if a
smaller file was written in place of an older larger one, and this put
garbage data at the end of the PNG.

On a suggestion from someone else writing a PNG reading lib, I updated
the test data to the latest PNG Suite, and now pngcs passes all of those
test images (that aren't supposed to fail). Some additional tests aren't
currently run because the Suite data no longer separates interlaced from
non-interlaced source images, or true-color from paletted versions of
the same image.

This also incorporates a PR submitted to leonbloy's version on GitHub
(who is the original author, but that repo was imported from Google Code
after I had already brought the same repo over to GitHub, and made some
changes).